scripts/set_sms_keywords/requirements.txt
boto3==1.34.25
python-dotenv==1.0.1
